requirements.txt
google-cloud-kms==2.12.0
crcmod